Extract dashboard mission control presenter (#269)
