+++ plugins/codex-autoresearch/lib/dashboard-view-model.ts
@@ -2,6 +2,7 @@ import { STATUS_VALUES, buildDecisionEnvelope, finiteMetric } from "./session-co
 import { redactEvidenceObject } from "./evidence-redaction.js";
 import { acceptedCurrentRuns, buildEvidenceRegistry } from "./evidence-registry.js";
 import { buildAiSummary } from "./dashboard-view-model/ai-summary.js";
+import { buildMissionControl } from "./dashboard-view-model/mission-control.js";
 import { buildWatchdogSummary } from "./dashboard-view-model/watchdog-summary.js";
 import {
   actionMetadataForKind,
@@ -16,6 +17,7 @@ import {
 import type { DashboardContext } from "../dashboard/src/types.js";
 
 export { buildAiSummary } from "./dashboard-view-model/ai-summary.js";
+export { buildMissionControl } from "./dashboard-view-model/mission-control.js";
 export { buildWatchdogSummary } from "./dashboard-view-model/watchdog-summary.js";
 
 type LooseObject = Record<string, any>;
@@ -2238,147 +2240,6 @@ function decisionEnvelopeUtility(kind: string): string {
     return "Finalization is ready after higher-priority loop checks.";
   if (kind === "baseline") return "Establish the benchmark floor before tuning.";
   return "Decision envelope is the authoritative next-action source.";
-}
-
-export function buildMissionControl({
-  current,
-  setupPlan,
-  guidedSetup,
-  qualityGap,
-  finalizePreview,
-  experimentMemory,
-  actionRail,
-  commands,
-}: LooseObject) {
-  const commandMap = commandLookup(commands);
-  const stage = guidedSetup?.stage || "ready";
-  const lastRun = guidedSetup?.lastRun || null;
-  const allowedStatuses = Array.isArray(lastRun?.allowedStatuses) ? lastRun.allowedStatuses : [];
-  const suggestedStatus =
-    lastRun?.safeSuggestedStatus ||
-    lastRun?.suggestedStatus ||
-    (allowedStatuses.length === 1 ? allowedStatuses[0] : "");
-  const hasFreshLastRun = Boolean(lastRun && lastRun?.freshness?.fresh !== false);
-  const canLog = stage === "needs-log-decision" && hasFreshLastRun && allowedStatuses.length > 0;
-  const qualityGapOpen = Number(qualityGap?.open);
-  const hasQualityGaps = Number.isFinite(qualityGapOpen) && qualityGapOpen > 0;
-  const setupState =
-    stage === "needs-setup" ? "ready" : setupPlan?.configured || current.length ? "done" : "idle";
-  const gapState = qualityGap ? (hasQualityGaps ? "ready" : "done") : "idle";
-  const logState = lastRun ? (hasFreshLastRun ? "ready" : "blocked") : "idle";
-  const finalizeState = finalizePreview?.ready ? "ready" : current.length ? "idle" : "blocked";
-  const activeStep = canLog
-    ? "log"
-    : stage === "needs-setup"
-      ? "setup"
-      : hasQualityGaps
-        ? "gaps"
-        : finalizePreview?.ready
-          ? "finalize"
-          : qualityGap
-            ? "gaps"
-            : actionRail?.[0]?.kind || "next";
-  return {
-    activeStep,
-    staticFallback: "Serve the dashboard locally for a fresh readout; use CLI for actions.",
-    steps: [
-      missionStep({
-        id: "setup",
-        title: "Setup",
-        state: setupState,
-        detail:
-          guidedSetup?.stage === "needs-setup"
-            ? guidedSetup.nextAction
-            : "Session setup is readable.",
-        safeAction: "setup-plan",
-        command: guidedSetup?.commands?.setup || commandMap.get("setup plan"),
-        commandLabel: "Setup",
-      }),
-      missionStep({
-        id: "gaps",
-        title: "Gap review",
-        state: gapState,
-        detail: qualityGap
-          ? `${qualityGap.open} open / ${qualityGap.total} total in ${qualityGap.slug}.`
-          : "No research gap file detected.",
-        safeAction: "gap-candidates",
-        command: commandMap.get("gap candidates"),
-        commandLabel: "Gaps",
-      }),
-      missionStep({
-        id: "log",
-        title: "Log decision",
-        state: logState,
-        detail: canLog
-          ? `Last packet is ready to log as ${suggestedStatus || "an allowed status"}.`
-          : lastRun?.freshness?.reason || "No fresh last-run packet is waiting.",
-      }),
-      missionStep({
-        id: "finalize",
-        title: "Finalize",
-        state: finalizeState,
-        detail:
-          finalizePreview?.nextAction || "Preview review branches after kept evidence is ready.",
-        safeAction: "finalize-preview",
-        command: commandMap.get("finalize preview"),
-        commandLabel: "Preview",
-      }),
-    ],
-    logDecision: {
-      available: canLog,
-      allowedStatuses,
-      suggestedStatus,
-      metric: lastRun?.metric ?? null,
-      lastRunFingerprint: lastRun?.fingerprint || "",
-      statusGuidance: lastRun?.statusGuidance || "",
-      defaultDescription:
-        suggestedStatus === "discard"
-          ? "Describe the discarded packet"
-          : suggestedStatus === "checks_failed"
-            ? "Describe the failed checks"
-            : "Describe the kept change",
-      asiTemplate: lastRun?.asiTemplate || {},
-      requiresDescription: true,
-      requiresConfirmation: true,
-    },
-    nextAction:
-      actionRail?.[0]?.detail ||
-      experimentMemory?.latestNextAction ||
-      guidedSetup?.nextAction ||
-      "",
-  };
-}
-
-function missionStep({
-  id,
-  title,
-  state,
-  detail,
-  safeAction = "",
-  command = "",
-  commandLabel = "Copy read-only command",
-  mutates = false,
-}: {
-  id: string;
-  title: string;
-  state: string;
-  detail: string;
-  safeAction?: string;
-  command?: string;
-  commandLabel?: string;
-  mutates?: boolean;
-}) {
-  const safeCommand = dashboardReadOnlyCommand(command);
-  return {
-    id,
-    title,
-    state,
-    detail,
-    safeAction,
-    command: safeCommand,
-    primaryCommand: safeCommand ? { label: commandLabel, command: safeCommand } : null,
-    mutates,
-  };
 }
 
 function actionItem({
